pyproject.toml
[tool.poetry]
name = "vsb-discord-bot"
version = "0.1.0"
description = "Discord moderation and community management bot with AI integration, advanced moderation, economy, and automation."
authors = ["VSB Discord Bot Team <dev@vsb-bot.example.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/KasheK420/py_VSB_Discord_Moderation_BOT"
repository = "https://github.com/KasheK420/py_VSB_Discord_Moderation_BOT"
keywords = ["discord", "bot", "moderation", "ai", "community", "postgresql", "docker"]
packages = [{ include = "bot" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

# runtime
"discord.py" = "^2.3.0"
aiohttp = "^3.9.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"
alembic = "^1.13.0"
sqlalchemy = { version = "^2.0.0", extras = ["asyncio"] }
psycopg2-binary = "^2.9.0"
pillow = "^10.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-alembic = "^0.10.0"

flake8 = "^6.1.0"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.0"
ruff = "^0.6.5"
pylint = "^2.17.0"

ipython = "^8.14.0"
ipdb = "^0.13.13"
memory-profiler = "^0.61.0"
line-profiler = "^4.1.0"
psutil = "^5.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ---- Tool Configs ----

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["bot", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E","F","I","UP","B","C4","SIM","RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
plugins = []
